updated styling
changed title of profile projects section to be either the display name or username.
changed the explore button on the home page to scroll to the projects section.
changed behavior of the scroll to top to be smooth instead of instant.
changed styling to fix some issues between light and dark